requirements.txt
anthropic==0.32.0
openai-whisper==20231117
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
pydantic==2.7.4
jinja2==3.1.4
